add objective.ts file
